+++ assets/js/yaml/util.ts
@@ -1,20 +1,28 @@
-import type {
-  WorkflowState,
-  SpecJob,
-  SpecTrigger,
-  SpecEdge,
-  StateJob,
-  StateTrigger,
-  StateEdge,
-  WorkflowSpec,
-} from './types';
+import Ajv, { type ErrorObject } from 'ajv';
+import YAML from 'yaml';
 import { randomUUID } from '../common';
 import workflowV1Schema from './schema/workflow-spec.json';
-import YAML from 'yaml';
-import Ajv, { type ErrorObject } from 'ajv';
+import type {
+  Position,
+  SpecEdge,
+  SpecJob,
+  SpecTrigger,
+  StateEdge,
+  StateJob,
+  StateTrigger,
+  WorkflowSpec,
+  WorkflowState,
+} from './types';
 
 const hyphenate = (str: string) => {
   return str.replace(/\s+/g, '-');
+};
+
+const roundPosition = (pos: Position): Position => {
+  return {
+    x: Math.round(pos.x),
+    y: Math.round(pos.y),
+  };
 };
 
 export const convertWorkflowStateToSpec = (
@@ -26,17 +34,26 @@ export const convertWorkflowStateToSpec = (
     jobDetails.name = job.name;
     jobDetails.adaptor = job.adaptor;
     jobDetails.body = job.body;
+
+    const jobPos = workflowState.positions?.[job.id];
+    if (jobPos) {
+      jobDetails.pos = roundPosition(jobPos);
+    }
     jobs[hyphenate(job.name)] = jobDetails;
   });
 
   const triggers: { [key: string]: SpecTrigger } = {};
   workflowState.triggers.forEach(trigger => {
-    const triggerDetails: SpecTrigger = {} as SpecTrigger;
-    triggerDetails.type = trigger.type;
+    const triggerDetails: SpecTrigger = { type: trigger.type } as SpecTrigger;
     if (trigger.cron_expression) {
       triggerDetails.cron_expression = trigger.cron_expression;
     }
     triggerDetails.enabled = trigger.enabled;
+
+    const triggerPos = workflowState.positions?.[trigger.id];
+    if (triggerDetails.type !== 'kafka' && triggerPos) {
+      triggerDetails.pos = roundPosition(triggerPos);
+    }
     // TODO: handle kafka config
     triggers[trigger.type] = triggerDetails;
   });
@@ -88,27 +105,36 @@ export const convertWorkflowStateToSpec = (
 export const convertWorkflowSpecToState = (
   workflowSpec: WorkflowSpec
 ): WorkflowState => {
+  const positions: Record<string, Position> = {};
   const stateJobs: Record<string, StateJob> = {};
   Object.entries(workflowSpec.jobs).forEach(([key, specJob]) => {
+    const uId = randomUUID();
     stateJobs[key] = {
-      id: randomUUID(),
+      id: uId,
       name: specJob.name,
       adaptor: specJob.adaptor,
       body: specJob.body,
     };
+    if (specJob.pos) positions[uId] = specJob.pos;
   });
 
   const stateTriggers: Record<string, StateTrigger> = {};
   Object.entries(workflowSpec.triggers).forEach(([key, specTrigger]) => {
+    const uId = randomUUID();
     const trigger = {
-      id: randomUUID(),
+      id: uId,
       type: specTrigger.type,
       enabled: true,
     };
 
-    if (specTrigger.type === 'cron') {
-      trigger.cron_expression = specTrigger.cron_expression;
+    if (specTrigger.type !== 'kafka' && specTrigger.pos) {
+      positions[uId] = specTrigger.pos;
     }
+
+    if (specTrigger)
+      if (specTrigger.type === 'cron') {
+        trigger.cron_expression = specTrigger.cron_expression;
+      }
 
     // TODO: handle kafka config
 
@@ -168,6 +194,7 @@ export const convertWorkflowSpecToState = (
     jobs: Object.values(stateJobs),
     edges: Object.values(stateEdges),
     triggers: Object.values(stateTriggers),
+    positions: Object.keys(positions).length ? positions : null, // null here is super important - don't mess with it
   };
 
   return workflowState;