+++ CHANGELOG.md
@@ -17,9 +17,24 @@ and this project adheres to
 
 ### Added
 
+- Allow users to edit position of nodes in the workflow
+  [#3123](https://github.com/OpenFn/lightning/issues/3123)
+- Minimap for easier workflow navigation
+  [#3125](https://github.com/OpenFn/lightning/issues/3125)
+- Added icons to control layout in the workflow
+  [PR #3242](https://github.com/OpenFn/lightning/pull/3242)
+
 ### Changed
 
+- Update React Flow to version 12
+  [PR #3242](https://github.com/OpenFn/lightning/pull/3242)
+- Create nodes and edges with the same button in the workflow
+  [#2175](https://github.com/OpenFn/lightning/issues/2175)
+
 ### Fixed
+
+- AI Assistant fails to send job context in subsequent messages
+[#3329](https://github.com/OpenFn/lightning/issues/3329)
 
 ## [v2.13.2] - 2025-06-18
 
@@ -55,6 +70,8 @@ This bug was introduced in version `v2.12.3-pre` on May 29th. If you're tracking
 
 - Report AI Assistant errors to Sentry
   [#3010](https://github.com/OpenFn/lightning/issues/3010)
+- Do not validate edge js condition expression
+  [#3028](https://github.com/OpenFn/lightning/issues/3028)
 
 ### Fixed